Split the files
